app/requirements.txt
fastapi==0.96.1
numpy==1.22.4
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
opencv_python_headless==4.7.0.72
pandas==1.5.3
Pillow==8.4.0
Pillow==9.5.0
pydantic==1.10.7
torch==2.0.1+cu118
torchvision==0.15.2+cu118
ultralytics==8.0.115
uvicorn==0.22.0
